backend/requirements.txt
Flask==3.0.0
flask-cors==4.0.0
flask-jwt-extended==4.6.0
python-dotenv==1.0.0
pymongo==4.6.1
bcrypt==4.1.2
Pillow==10.2.0
numpy==1.26.3
scikit-learn==1.4.0
opencv-python==4.9.0.80
torch==2.1.2
torchvision==0.16.2
facenet-pytorch==2.5.3
insightface==0.7.3
onnxruntime==1.16.3
pandas==2.1.4
tqdm==4.66.1

